🐛 [IOT] facilities 페이지 useSearchParams Suspense 래핑 (빌드 에러 수정)
